feat(cart): render the cart page
This feature renders the shopping cart page, where users can view the items
they've added to their cart, the individual amount to pay considering
the number of times an item appears in the cart, and the total payable
amount. Clients can also delete a group of variations from their cart.

- Implements the `AddToCartView` to store product variations in the
user's session-based shopping cart.
- Implements the `RemoveFromCartView` to remove clusters of variations
from the cart.
- Creates the `cart` template to render information about the items
selected.
- Creates the utilities `cart_totals` and `cart_total_qty` to provide
the total sum of the purchase and the total number of items in each
variation cluster stored in the cart.
- Creates tailored `cart_total_qty` and `cart_totals` filters for 
template formatting.
- Updates the `cart` icon in the navbar according to the current
client's session.
- Fixes the `cart` navbar rendering issue.
- Creates empty templates for further implementation.
- Creates an empty form for the `profiles` app.